Put movement code in its own file
