+++ src/main.cpp
@@ -2,13 +2,13 @@
 
 #include <chrono>
 #include <iostream>
-#include <math.h>
 #include <thread>
 #include <vector>
 
 #include <glad/glad.h>
 #include <GLFW/glfw3.h>
 
+#include <game/movement.hpp>
 #include <game/player.hpp>
 #include <graphics/shader.hpp>
 #include <graphics/texture.hpp>
@@ -16,64 +16,6 @@
 namespace sniper {
 
 GLFWwindow * wn;
-
-/* Maximum possible speeds for the player. */
-float MAX_X = 0.5f, MAX_Y = 0.5f;
-
-/* How slowly (the inverse of how fast in the internal calculation) the player's
-speed updates to the maximum it can be in the direction they are traveling. */
-float SLIPPERINESS = 25;
-
-/* How much friction there is. */
-float FRICTION = 0.01f;
-
-/* FPS. */
-int FPS = 50;
-
-/* Get the new X and Y speed of the player, given the current X and Y speed and
-readings from GLFW of the currently pressed keys. */
-std::pair<float, float> new_speed(std::pair<float, float> old_speed) {
-
-    std::pair<float, float> ret = old_speed;
-
-    // For horizontal and vertical axes, we return -1, 0, or 1 -- which
-    // direction in each to increase speed in.
-    int hinc = 0, vinc = 0;
-
-    if (glfwGetKey(wn, GLFW_KEY_A) == GLFW_PRESS) {
-        hinc -= 1;
-    }
-    if (glfwGetKey(wn, GLFW_KEY_D) == GLFW_PRESS) {
-        hinc += 1;
-    }
-    if (glfwGetKey(wn, GLFW_KEY_W) == GLFW_PRESS) {
-        vinc += 1;
-    }
-    if (glfwGetKey(wn, GLFW_KEY_S) == GLFW_PRESS) {
-        vinc -= 1;
-    }
-
-    /* We slow down a tiny bit if nothing is pressed. */
-    if (hinc == 0 && vinc == 0) {
-        ret.first  *= (1.0f - FRICTION);
-        ret.second *= (1.0f - FRICTION);
-        return ret;
-    }
-
-    // The division by the sum is so that the maximum TOTAL velocity is the same
-    // even if both horizontal and vertical velocity components are changing.
-    float div = sqrt(hinc * hinc + vinc * vinc);
-    float ideal_x = MAX_X * hinc / (div * FPS);
-    float ideal_y = MAX_Y * vinc / (div * FPS);
-
-    float update_frac = 1 / SLIPPERINESS;
-
-    ret.first  = ret.first  * (1 - update_frac) + ideal_x * update_frac;
-    ret.second = ret.second * (1 - update_frac) + ideal_y * update_frac;
-
-    return ret;
-
-}
 
 void init() {
 
@@ -110,7 +52,7 @@ void mainloop() {
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
         // Update the player speed and position.
-        player.setvel(new_speed(player.getvel()));
+        player.setvel(mvmt::new_speed(wn, player.getvel()));
         player.do_pos_updates();
 
         player.do_bounce(0.9f);
@@ -120,7 +62,7 @@ void mainloop() {
 		glfwSwapBuffers(wn);
 		glfwPollEvents();
 
-		std::this_thread::sleep_until(start_of_frame + std::chrono::milliseconds(1000 / FPS));
+		std::this_thread::sleep_until(start_of_frame + std::chrono::milliseconds(1000 / mvmt::FPS));
 
 	}
